+++ GiniVision/Classes/Core/Extensions/UIViewController.swift
@@ -76,9 +76,6 @@ extension UIViewController {
         
         let alertViewController = UIAlertController(title: nil, message: message, preferredStyle: .alert)
         
-        // Ignore dark mode
-        alertViewController.useLightUserInterfaceStyle()
-        
         alertViewController.addAction(UIAlertAction(title: cancelActionTitle,
                                                     style: .cancel,
                                                     handler: { _ in
@@ -94,11 +91,5 @@ extension UIViewController {
         }
         
         return alertViewController
-    }
-    
-    func useLightUserInterfaceStyle() {
-        if #available(iOS 13.0, *) {
-            overrideUserInterfaceStyle = .light
-        }
     }
 }