+++ GiniVision/Classes/Core/Screens/Analysis/AnalysisViewController.swift
@@ -160,12 +160,6 @@ import UIKit
         addErrorView()
     }
     
-    public override func viewDidLoad() {
-        super.viewDidLoad()
-        // Ignore dark mode
-        useLightUserInterfaceStyle()
-    }
-    
     override public func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         didShowAnalysis?()