+++ GiniVision/Classes/Core/Screens/Document picker/DocumentPickerCoordinator.swift
@@ -175,9 +175,6 @@ public final class DocumentPickerCoordinator: NSObject {
                                    device: UIDevice = UIDevice.current) {
         let documentPicker = UIDocumentPickerViewController(documentTypes: acceptedDocumentTypes, in: .import)
         
-        // Ignore dark mode
-        documentPicker.useLightUserInterfaceStyle()
-        
         documentPicker.delegate = self
         
         if #available(iOS 11.0, *) {